refactor autograd_cache test
